pyproject.toml
[tool.poetry]
name = "handwriting_ocr"
version = "0.0.1"
description = "Handwriting recognition demo app"
authors = ["Jiayu Liu <jiayu@blacklake.cn>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
streamlit = "^1.26.0"
watchdog = "^3.0.0"
pydantic = "^2.3.0"

# [tool.poetry.scripts]
# serve = ":main"

[tool.poetry.group.dev.dependencies]
pytest-black = "^0.3.12"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
